# getting-started/src/main/resources/META-INF/openapi.yml
---
openapi: 3.1.3
info:
  title: Some static API
  description: Our test API description
  version: "1.2"
servers:
- url: //quarkus.study.local
  description: Dev server
paths:
  /book:
    get:
      tags:
      - Book Resource
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      tags:
      - Book Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        "200":
          description: OK
  /book/{id}:
    put:
      tags:
      - Book Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    delete:
      tags:
      - Book Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /hello:
    get:
      tags:
      - Greeting Resource
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /hello/{name}:
    get:
      tags:
      - Greeting Resource
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    Book:
      required:
      - name
      - author
      type: object
      properties:
        name:
          pattern: \S
          type: string
          nullable: false
        author:
          pattern: \S
          type: string
          nullable: false
        pages:
          format: int32
          minimum: 1
          type: integer